v0.0.7 - refactor graphql schema builder
